+++ demo2/android/app/src/main/java/com/demo2/demo2/MainActivity.java
@@ -1,0 +1,6 @@
+package com.demo2.demo2;
+
+import io.flutter.embedding.android.FlutterActivity;
+
+public class MainActivity extends FlutterActivity {
+}