+++ demo2/lib/app/modules/home/bindings/home_binding.dart
@@ -1,0 +1,14 @@
+import 'package:refreshed/refreshed.dart';
+
+import '../controllers/home_controller.dart';
+
+class HomeBinding extends Binding {
+  @override
+  List<Bind> dependencies() {
+    return [
+      Bind.lazyPut<HomeController>(
+        () => HomeController(),
+      ),
+    ];
+  }
+}