+++ demo2/lib/app/modules/home/controllers/home_controller.dart
@@ -1,0 +1,11 @@
+import 'package:refreshed/refreshed.dart';
+
+class HomeController extends GetxController {
+  final count = 0.obs;
+  @override
+  void onInit() {
+    super.onInit();
+  }
+
+  void increment() => count.value++;
+}