+++ demo2/lib/app/modules/profile/bindings/profile_binding.dart
@@ -1,0 +1,14 @@
+import 'package:refreshed/refreshed.dart';
+
+import '../controllers/profile_controller.dart';
+
+class ProfileBinding extends Binding {
+  @override
+  List<Bind> dependencies() {
+    return [
+      Bind.lazyPut<ProfileController>(
+        () => ProfileController(),
+      ),
+    ];
+  }
+}