+++ demo2/lib/app/modules/profile/controllers/profile_controller.dart
@@ -1,0 +1,11 @@
+import 'package:refreshed/refreshed.dart';
+
+class ProfileController extends GetxController {
+  final count = 0.obs;
+  @override
+  void onInit() {
+    super.onInit();
+  }
+
+  void increment() => count.value++;
+}