+++ demo2/lib/app/modules/profile/views/profile_view.dart
@@ -1,0 +1,24 @@
+import 'package:flutter/material.dart';
+
+import 'package:refreshed/refreshed.dart';
+
+import '../controllers/profile_controller.dart';
+
+class ProfileView extends GetView<ProfileController> {
+  const ProfileView({super.key});
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('ProfileView'),
+        centerTitle: true,
+      ),
+      body: const Center(
+        child: Text(
+          'ProfileView is working',
+          style: TextStyle(fontSize: 20),
+        ),
+      ),
+    );
+  }
+}