+++ demo2/lib/app/routes/app_routes.dart
@@ -1,0 +1,8 @@
+part of 'app_pages.dart';
+// DO NOT EDIT. This is code generated via package:get_cli/get_cli.dart
+
+abstract class Routes {
+  Routes._();
+  static const home = '/home';
+  static const profile = '/profile';
+}