+++ demo2/lib/main.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+import 'package:refreshed/refreshed.dart';
+
+import 'app/routes/app_pages.dart';
+
+void main() {
+  runApp(
+    GetMaterialApp(
+      title: "Application",
+      initialRoute: AppPages.initial,
+      getPages: AppPages.routes,
+    ),
+  );
+}